pyproject.toml
[project]
name = "mydjango-crm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
classifiers = [ "Programming Language :: Python :: 3 :: Only" ]
dependencies = [
  "certifi>=2025.11.12",
  "dj-database-url>=3.0.1",
  "django>=5.2,<6",
  "django-googler>=0.0.20",
  "django-timescaledb>=0.2.13",
  "google-api-python-client>=2.187.0",
  "pre-commit>=4.4",
  "psycopg[binary]>=3.2.12",
  "python-dotenv>=1.2.1",
  "requests>=2.32.5",
  "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
  "pip>=25.3",
  "rav>=0.1.1",
]
